fix: appRouter의 디렉토리 구조를 명세에 맞게 변경
