+++ src/features/carts/hooks/useAddCartItemQuery.ts
@@ -16,9 +16,9 @@ function useAddCartItemQuery() {
         queryClient.cancelQueries(getCartListQuery.queryKey());
 
         // 이전 장바구니 목록 스냅샷을 가져옵니다.
-        const previousCartListData = queryClient.getQueryData(
+        const previousCartListData = queryClient.getQueryData<GetCartListQueryFnData>(
           getCartListQuery.queryKey(),
-        ) as GetCartListQueryFnData | undefined;
+        );
 
         // 스냅샷이 존재하는 경우 Optimistic 업데이트합니다.
         if (previousCartListData) {
